requirements.txt
streamlit
pandas
cohere
python-dotenv
scikit-learn
matplotlib
